Introduce `test` logger fn
