+++ rust/connlib/snownet/tests/lib.rs
@@ -21,8 +21,7 @@ fn connection_times_out_after_20_seconds() {
 
 #[test]
 fn connection_without_candidates_times_out_after_10_seconds() {
-    let _ = tracing_subscriber::fmt().with_test_writer().try_init();
-
+    let _guard = firezone_logging::test("trace");
     let start = Instant::now();
 
     let (mut alice, mut bob) = alice_and_bob();
@@ -38,8 +37,7 @@ fn connection_without_candidates_times_out_after_10_seconds() {
 
 #[test]
 fn connection_with_candidates_does_not_time_out_after_10_seconds() {
-    let _ = tracing_subscriber::fmt().with_test_writer().try_init();
-
+    let _guard = firezone_logging::test("trace");
     let start = Instant::now();
 
     let (mut alice, mut bob) = alice_and_bob();
