+++ rust/connlib/tunnel/Cargo.toml
@@ -42,6 +42,7 @@ serde_json = "1.0"
 test-strategy = "0.3.1"
 tracing-appender = "0.2.3"
 tracing-subscriber = { version = "0.3", features = ["env-filter"] }
+firezone-logging = { workspace = true }
 
 [features]
 proptest = ["dep:proptest", "connlib-shared/proptest"]
