+++ rust/headless-client/Cargo.toml
@@ -14,6 +14,7 @@ clap = { version = "4.5", features = ["derive",  "env", "string"] }
 connlib-client-shared = { workspace = true }
 connlib-shared = { workspace = true }
 firezone-bin-shared = { workspace = true }
+firezone-logging = { workspace = true }
 futures = "0.3.30"
 humantime = "2.1"
 ip_network = { version = "0.4", default-features = false }
