+++ rust/headless-client/src/main.rs
@@ -8,7 +8,7 @@ use connlib_shared::{get_user_agent, DEFAULT_MTU};
 use firezone_bin_shared::{
     new_dns_notifier, new_network_notifier,
     platform::{tcp_socket_factory, udp_socket_factory},
-    setup_global_subscriber, TunDeviceManager, TOKEN_ENV_KEY,
+    TunDeviceManager, TOKEN_ENV_KEY,
 };
 use firezone_headless_client::{
     device_id, signals, CallbackHandler, CliCommon, DnsController, InternalServerMsg, IpcServerMsg,
@@ -130,7 +130,7 @@ fn main() -> Result<()> {
         .as_deref()
         .map(file_logger::layer)
         .unzip();
-    setup_global_subscriber(layer);
+    firezone_logging::setup_global_subscriber(layer);
 
     tracing::info!(
         arch = std::env::consts::ARCH,
